+++ src/utils/loadDeck.js
@@ -1,0 +1,109 @@
+fn.deckLoader = (() => {
+  const actions = ['addCard', 'removeCard', 'removeAllCards', 'clearArtifacts', 'addArtifact'];
+  const mockLastOpenedDialog = { close() {} };
+  const pending = [];
+  let templastOpenedDialog;
+
+  function clear() {
+    pending.splice(0, pending.length);
+  }
+
+  function processNext() {
+    if (global('lastOpenedDialog') === mockLastOpenedDialog) {
+      globalSet('lastOpenedDialog', templastOpenedDialog);
+    }
+    const job = pending.shift();
+    if (!job) return;
+
+    if (job.action === 'validate') {
+      loadDeck({ artifacts: job.artifacts, cards: job.cards });
+    } else if (job.action === 'clear') {
+      global('removeAllCards')();
+    } else if (job.action === 'remove') {
+      global('removeCard')(parseInt(job.id, 10), job.shiny === true);
+    } else if (job.action === 'clearArtifacts') {
+      global('clearArtifacts')();
+    } else if (job.action === 'addArtifact') {
+      templastOpenedDialog = global('lastOpenedDialog');
+      globalSet('lastOpenedDialog', mockLastOpenedDialog);
+      global('addArtifact')(job.id);
+    } else {
+      global('addCard')(parseInt(job.id, 10), job.shiny === true);
+    }
+  }
+
+  eventManager.on('Deck:postChange', ({ action, data }) => {
+    if (!actions.includes(action) || !data) return;
+    if (data.status === 'error') {
+      clear();
+    } else {
+      processNext();
+    }
+  });
+
+  function loadDeck({ cards = [], artifacts = [] }) {
+    clear();
+    if (cards.length > 25 || artifacts > 2) return;
+    // Get current deck
+    const cDeck = global('decks')[global('soul')];
+    // Remove unwanted cards
+    if (cDeck.length) {
+      const deck = [...cards];
+      const removals = cDeck.map(({ id, shiny }) => ({
+        id,
+        shiny,
+        action: 'remove',
+      })).filter((card) => !deck.some((card2, ind) => {
+        const found = card2.id === card.id && (card.shiny || false) === (card2.shiny || false);
+        if (found) { // Remove the item
+          deck.splice(ind, 1);
+        }
+        return found;
+      }));
+
+      if (removals.length > 13 || removals.length === cDeck.length) { // Too much to do (Cards in deck + 1) OR removing all cards
+        pending.push({ action: 'clear' });
+        pending.push(...cards); // Add everything
+      } else { // Remove bad cards
+        pending.push(...removals);
+        pending.push(...deck); // Add missing cards
+      }
+    } else { // All new cards
+      pending.push(...cards);
+    }
+    const cArts = global('decksArtifacts')[global('soul')];
+    if (!cArts.every((art) => artifacts.includes(art) || artifacts.includes(art.id))) { // Clear artifacts (if they don't match)
+      pending.push({ action: 'clearArtifacts' });
+      artifacts.forEach((art) => {
+        pending.push({
+          id: art.id || art,
+          action: 'addArtifact',
+        });
+      });
+    } else { // Partial or empty
+      const ids = cArts.map((art) => art.id);
+      artifacts.forEach((art) => {
+        const id = art.id || art;
+        if (ids.includes(id)) return;
+        pending.push({
+          id,
+          action: 'addArtifact',
+        });
+      });
+    }
+    // Validate
+    if (pending.length) {
+      pending.push({
+        action: 'validate',
+        cards: cards.map(({ id, shiny }) => ({ id, shiny })),
+        artifacts: artifacts.map((art) => art.id || art),
+      });
+      processNext();
+    }
+  }
+
+  return Object.freeze({
+    load: onPage('Decks') ? loadDeck : noop,
+    clear,
+  });
+})();