+++ changelog.md
@@ -1,6 +1,6 @@
 # UnderScript Changelog
 
-## Version 0.45.0 (2021-06-06)
+## Version 0.45.0 (2021-06-16)
 ### New Features
 1. Added ability to buy multiple packs
 1. Added confirmation for buying packs with UCP
@@ -8,6 +8,7 @@
 1. Added a random avatar button
 1. Added setting to prefer shiny on import (enabled by default)
 1. Added "fill deck" button
+1. Added confirmation for buying cosmetics
 ### Fixes
 1. Fixed bug with base card skin setting
 ### Plugins